feat: support ip streamdecks (#26)
