+++ .nvmrc
@@ -1,1 +1,1 @@
-18.19
+22.15.1
