+++ packages/input-manager/src/integrations/streamdeck/$schemas/options.json
@@ -19,6 +19,19 @@
 			"ui:title": "Device Index",
 			"ui:description": "The index on the list of attached Stream Deck devices"
 		},
+		"ip": {
+			"type": "string",
+			"ui:title": "IP Address",
+			"ui:description": "The ip address of a network connected Stream Deck"
+		},
+		"port": {
+			"type": "number",
+			"ui:title": "IP Address",
+			"ui:description": "The port number of a network connected Stream Deck. Only used if the device is connected over IP",
+			"default": 5343,
+			"min": 1,
+			"max": 65535
+		},
 		"brightness": {
 			"type": "integer",
 			"minimum": 0,
@@ -83,10 +96,7 @@
 						"fontWeight": {
 							"type": "string",
 							"ui:title": "Label Font Bold Weight",
-							"enum": [
-								"normal",
-								"bold"
-							],
+							"enum": ["normal", "bold"],
 							"default": "normal"
 						},
 						"fontWidth": {
@@ -142,12 +152,7 @@
 							"type": "string",
 							"ui:title": "Capitalize text",
 							"ui:description": "Transform label text capitalization",
-							"enum": [
-								"none",
-								"capitalize",
-								"uppercase",
-								"lowercase"
-							],
+							"enum": ["none", "capitalize", "uppercase", "lowercase"],
 							"default": "none"
 						},
 						"lineClamp": {
@@ -169,9 +174,7 @@
 							"default": ""
 						}
 					},
-					"required": [
-						"id"
-					],
+					"required": ["id"],
 					"additionalProperties": false
 				}
 			},
